+++ frontmatter.json
@@ -1,0 +1,3 @@
+{
+  "$schema": "https://frontmatter.codes/frontmatter.schema.json"
+}
