+++ src/App.js
@@ -3,11 +3,11 @@ import Router from './router/Router';
 
 
 function App() {
-  return (
-    <div className="App">
-     <Router/>
-    </div>
-  );
+    return (
+        <div className="App">
+            <Router/>
+        </div>
+    );
 }
 
 export default App;
