+++ src/components/downloads/Downloads.jsx
@@ -1,14 +1,16 @@
 import './downloads.css'
-import DownloadPersonal from './DownloadPersonal';
-import DownloadWorkshops from './DownloadWorkshops';
-import DownloadDocuments from './DownloadDocuments';
-import DownloadHolidays from './DownloadHolidays';
-import DownloadSignin from './DownloadSignin';
+import DownloadPersonal from './PDF/DownloadPersonal';
+import DownloadWorkshops from './PDF/DownloadWorkshops';
+import DownloadDocuments from './PDF/DownloadDocuments';
+import DownloadHolidays from './PDF/DownloadHolidays';
+import DownloadSignin from './PDF/DownloadSignin';
+import ExcelPersonal from './Excel/ExcelPersonal'
 
 function Downloads() {
     return (
         <div className="downloads">
             <DownloadPersonal/>
+            <ExcelPersonal/>
             <DownloadDocuments/>
             <DownloadWorkshops/>
             <DownloadHolidays/>
