+++ src/components/downloads/PDF/DownloadSignin.jsx
@@ -1,13 +1,13 @@
 import React, { useState, useEffect } from 'react';
 import html2canvas from 'html2canvas';
 import { jsPDF } from 'jspdf';
-import './downloads.css'
+import '../downloads.css'
 
-function DownloadDocuments() {
-    const [content, setContent] = useState('');
+function DownloadSignin() {
+    const [content, setContent] = useState([]);
 
     useEffect(() => {
-        fetch('http://127.0.0.1:8000/apidocuments/list')
+        fetch('http://127.0.0.1:8000/apisignin/list')
             .then(response => response.json())
             .then(data => setContent(data))
             .catch(error => console.log(error));
@@ -16,22 +16,26 @@ function DownloadDocuments() {
     const downloadPdf = async () => {
         const tableHeader = `
           <tr>
-            <th style="font-size: 14px;">Document</th>
-            <th style="font-size: 14px;">Docname</th>
-            <th style="font-size: 14px;">Description</th>
+            <th style="font-size: 14px;">TimeStart</th>
+            <th style="font-size: 14px;">TimeStop</th>
+            <th style="font-size: 14px;">TimeRestart</th>
+            <th style="font-size: 14px;">TimeFinish</th>
+            <th style="font-size: 14px;">Hourcount</th>
             <th style="font-size: 14px;">User</th>
           </tr>
         `;
-      
-        const tableRows = content.map(documents => `
+
+        const tableRows = content.map(signin => `
           <tr>
-            <td><img src="${documents.document}" /></td>
-            <td>${documents.docname}</td>
-            <td>${documents.description}</td>
-            <td>${documents.personal}</td>
+            <td>${signin.timestart}</td>
+            <td>${signin.timestop}</td>
+            <td>${signin.timerestart}</td>
+            <td>${signin.timefinish}</td>
+            <td>${signin.hourcount}</td>
+            <td>${signin.personal}</td>
           </tr>
         `).join('');
-      
+
         // Add custom font faces and styles
         const html = `
           <html>
@@ -71,53 +75,33 @@ function DownloadDocuments() {
           </body>
           </html>
         `;
-      
+
         // Create a new element with the HTML content and append it to the document
         const element = document.createElement('div');
         element.innerHTML = html;
         document.body.appendChild(element);
-      
-        // Convert image URLs in the HTML to base64-encoded data URLs to embed the images in the HTML
-        const loadBase64Image = async (url) => {
-          const res = await fetch(url);
-          const blob = await res.blob();
-          const base64Data = await new Promise((resolve) => {
-            const reader = new FileReader();
-            reader.readAsDataURL(blob);
-            reader.onloadend = () => {
-              resolve(reader.result);
-            };
-          });
-          return base64Data;
-        };
-      
-        const images = element.querySelectorAll('img');
-      
-        try {
-          await Promise.all(Array.from(images).map(async (image) => {
-            const url = image.src;
-            const base64Data = await loadBase64Image(url);
-            image.setAttribute('src', base64Data);
-          }));
-        } catch (error) {
-          console.error('Error loading image:', error);
-        }
-      
+
         // Calculate the total width of the table
         const table = element.querySelector('table');
         const totalWidth = `${table.offsetWidth}px`;
-      
+
         // Set the width of the table to fit the content dynamically
         table.style.width = totalWidth;
-      
+
         // Convert HTML content to canvas using the new element
         try {
           const canvas = await html2canvas(element, {
             scrollY: -window.scrollY
           });
-          document.body.removeChild(element); // Remove the element from the document
-          const pdf = new jsPDF('p', 'mm', 'a4'); // Create portrait mode PDF
-          pdf.addImage(canvas.toDataURL('image/png'), 'PNG', 10, 10, 180, 0); // Set the width of the image based on the content automatically
+          // Remove the element from the document
+          document.body.removeChild(element);
+          // Create new image from canvas
+          const imgData = canvas.toDataURL('image/png');
+          // Create new PDF document
+          const pdf = new jsPDF('p', 'mm', 'a4');
+          // Set margins and add image to the PDF document
+          pdf.addImage(imgData, 'PNG', 10, 10, 180, 0);
+          // Save the PDF document
           pdf.save('download.pdf');
         } catch (error) {
           console.error('Error generating PDF:', error);
@@ -126,9 +110,9 @@ function DownloadDocuments() {
 
     return (
         <div>
-            <button className='btn' onClick={downloadPdf}>Download Documents PDF</button>
+            <button className='btn' onClick={downloadPdf}>Download Signin PDF</button>
         </div>
     );
 }
 
-export default DownloadDocuments;
+export default DownloadSignin;
