+++ src/pages/admin/Admin.jsx
@@ -5,9 +5,9 @@ import SelecAdmin from '../../components/selecusuario/selecAdmin/SelecAdmin'
 import NewUser from '../../components/create/NewUser'
 import NewTask from '../../components/create/NewTask'
 import NewWorkshop from '../../components/create/NewWorkshop'
-import DownloadDocuments from '../../components/downloads/DownloadDocuments'
-import DownloadHolidays from '../../components/downloads/DownloadHolidays'
-import DownloadSignin from '../../components/downloads/DownloadSignin'
+import Downloads from '../../components/downloads/Downloads'
+import Doc from '../../components/documents/Doc';
+
 
 function Admin() {
   return (
@@ -18,9 +18,8 @@ function Admin() {
       <NewUser />
       <NewTask />
       <NewWorkshop />
-      <DownloadSignin />
-      <DownloadDocuments />
-      <DownloadHolidays />
+      <Doc/>
+      <Downloads />
       
     </div>
   )
