+++ src/pages/login/login.css
@@ -1,45 +1,45 @@
-div{
+div {
     margin: 0 auto;
     padding: 0;
-    font-family: 'Poppins', sans-serif;
+    font-family: "Arial", sans-serif;
     box-sizing: border-box;
 }
 
-.desktop{
+.desktop {
     width: 100%;
 }
-.logo{
+.logo {
     width: 116px;
     margin-top: -90px;
     margin-bottom: 50px;
-    
+
 }
-.box-fichaje{
+.box-fichaje {
     display: flex;
     justify-content: center;
     align-items: center;
     margin-left: 500px;
-     
+
 }
-form{
+form {
     width: 700px;
     margin-bottom: 90px;
     margin-right: 150px;
     height: 80vh;
 }
-.inputContainer{
+.inputContainer {
     position: relative;
     height: 45px;
     width: 90%;
     margin-bottom: 17px;
 }
-.boxes{
+.boxes {
     position: absolute;
-    top: 0px;
+    top: 0;
     left: -150px;
     height: 100%;
     width: 100%;
-    border: 1px solid #DADCE0;
+    border: 1px solid #008894;
     border-radius: 7px;
     font-size: 16px;
     padding: 0 20px;
@@ -51,32 +51,32 @@ form{
     top: -7px;
     left: 3px;
     z-index: 10;
-    font-size: 14px;
+    font-size: 16px;
     font-weight: 600;
-    color: #1C1B1F;
-  }
+    color: #008894;
+}
 .boxes:focus {
-    border: 2px solid #1C1B1F;
-  }
+    border: 3px solid #A60032;
+}
 .boxes:not(:placeholder-shown)+ .subtitle {
     top: -7px;
     left: -140px;
     z-index: 10;
     font-size: 14px;
     font-weight: 600;
-  }
-  .subtitle{
+}
+.subtitle {
     position: absolute;
     top: 15px;
-    left: 0px;
+    left: 0;
     padding: 0 4px;
     background-color: white;
-    color: #DADCE0;
+    color: #A60032;
     font-size: 16px;
     transition: 0.5s;
     z-index: 0;
 }
-.btn{
+.btn {
     display: block;
     margin-left: auto;
     padding: 15px 30px;
@@ -93,36 +93,36 @@ form{
     background-color: #06c2d3;
     transform: translateY(-2px);
 }
-.personita{
+.personita {
     position: absolute;
     margin-top: 250px;
     margin-left: 350px;
     width: 470px;
 }
-@media screen and (max-width:767px) {
-    .logo{
+@media screen and(max-width:767px) {
+    .logo {
         width: 60px;
     }
-    .box-fichaje{
+    .box-fichaje {
         margin-left: 10px;
     }
-    form{
+    form {
         width: 200px;
-        margin-right:0px;
+        margin-right: 0;
         margin-left: 260px;
         margin-top: 30px;
     }
-    .btn{
+    .btn {
         margin-top: 80px;
         margin-left: -90px;
     }
-    .personita{
+    .personita {
         width: 300px;
         margin-left: 70px;
-        margin-right: 0px;
+        margin-right: 0;
         margin-top: 330px;
     }
-    picture{
+    picture {
         width: 100px;
     }
 }
