+++ src/pages/user/User.jsx
@@ -4,15 +4,19 @@ import SelecUsuario from '../../components/selecusuario/SelecUsuario'
 import Personal from '../../components/personal/Personal'
 import Signin from '../../components/signin/Signin'
 import Holidays from '../../components/holidays/Holidays'
-import DownloadPersonal from '../../components/downloads/DownloadPersonal'
+import Downloads from '../../components/downloads/Downloads'
+import Doc from '../../components/documents/Doc';
 
 function User() {
   return (
     <div>
       <NavbarUser />
+      <Signin/>
       <Personal />
       <SelecUsuario />
-      <DownloadPersonal />
+      <Holidays/>
+      <Doc/>
+      <Downloads />
 
     </div>
   )
